Loop restrictions hoisting
It is possible to restrict the range been traversed by an index in an IT loop
with accompanying conditional expressions. As IT loops may span into multiple
nested loops in C++, it is important to find out which restriction to apply to
what C++ loop to reduce iterations. This commit handles that issue.