Fixed a lexical analysis error
Previously if a code has comments just after the opening '{' anywhere, the
compiler was throwing syntax error. This seems to be complicated problem
to solve during parsing as we use new-lines as statement separators. Thus
as a simpler alternative, the lexical analyzer gobbles up all such comments
and their terminating new-lines.