Added a simplified implementation of reduction operations
Now a compilable IT source can have reduction operations. Currently,
reduction is implemented as sequential traversal of the the array or
arrays that are part of the reduction expression within a single PPU.
Ideally we would like to do reduction parallelly in multiple PPUs
then the result gets accumated according to the logic of the particular
operation. That is something we leave for the future.