Fixed an error in synchronization
Previous version was wrongly using the loop nesting information of the compute
stages that are source and sink of a synchronization to invoke signal and wait
routines on primitives as opposed to using that of the signaler and its counter-
part's, where actual interaction happens. This commit fixed that problem.