significant changes in the memory management code
That data partitioning is hierarchical in IT was not reflecting in allocation, part genaration,
and part management libraries in memory-management module. Consequently the code generation
library for elements of the former was erroneous too. This commit is the first step towards
correcting the error. It fixes the part-generation library and corresponding code generation
routines.