fixed a bunch of small runtime errors
We are still trying to run the matrix-matrix multiplication code for segmented memory
backends. The code is still failing as we need modifications in how we access data in
functions generated for compute stages based on the new memory allocation scheme. But
there were a lot of other errors that were throwing faults for LPU generation, input
output, and so on. May be there are more to come, but this commit fixed a lot of them.