Added code for interval description generation for hierarchical ranges
This will be integrated with part folding features in segmented memory back-end compiler
to generate interval descriptions for a segment's data content.

Note that the interval description generation process can incorporate several easy optimizations
that we should definitely do in the future to reduce the description length and process runtime.