update on some IT communicator settings
The up and down propagation sync communicators were getting stuck in the send and receive only
segments respectively at underlying MPI communicator setup time. This is because these two
communicator needs to know what other segments will be interacting with the segment the current
segments will be communicating to. This commit made changes so that relevant other segments' IDs
are available to the current segment to set up the MPI communicators properly.

Hopefully, this will work.