added a runtime counter for each communicator
This is needed to determine the current iteration numbe for send and receive request on the
communicator. Previously, we were using the repeat-iteration variable as the alternative.
That has the problem of getting reset within nested loops, rendering the communicator unusable
for future use. This commit should fix that problem.