fixed a bug in memory allocation
Previously, we were not checking if a segment has nothing to allocate for some data structures in
the computation. As a result, in several places we were creating objects and doing computations
that do not apply for empty (or partially empty) segments. This is a probable cause of the program
failure for block-LUF we have been experiencing in Rivana.

This commit fixes the empty segment problem.