fixed an MPI communicator setup error
Apparently every MPI process needs to participate in the MPI_Comm_create function. That
requires a complete understanding of which process should interect (or not interact) with
which others. This is a problematic scenario in IT where each segment only knows who else
should be interacting with it (if it will participate in communication) or if it is not
going to be concerned about a particular dependency.

Therefore, this commit changes communicator setup from using MPI_Comm_create to
MPI_Comm_split which does not require an uniform view of the whole system.

The code will still not run for small input where some segments do not participate in
communication as we haven't figured yet how to rule out non-participating segments from
scalar communicators.