a work in progress commit on supporting replication over down sync communicator
The updated code for block LU factorization involves a scenario of down sync communication
where the data is replicated on the sender side. Such scenarios were previously supported
in neither up nor down communicator. This commit attempted to make it work for the down sync
communicator. The code is, however, untested. Therefore it is a work in progress commit.

Also fixed a problem with redundant memory allocation. Hopefully, that change has no side
effect in the computation logic.