fixed a bug related to LPU construction
Previously variables used inside reduction expressions only in compute stages of an LPS
were not assigned in the LPUs of that LPS. Their references were NULL. This happened as
we wanted to treat reductions differently from read/write but latter did not account for
the distinction in the memory allocation and LPU generation strategy. This commit fixed
that problem.