hopefully fixed the error that caused the block LUF code to fail
The problem was related to synchronization. One thread in a segment was updating the data structure
and communicating the update to all other segments while its companions in the same segment might
still be needing the old value of the data structure for their computation. This commit ensures that
any update has been done after all local computations depending on the underlying data structure
has ended.

Hopefully, this will work.