A change in the logic of LPS assignment to repeat loops
This commit checks if a repeat loop embodies sub-flow that manipulates data structures
with overlapping partitions. For such repeat loops, the new logic tries to assign the
repeat loop a higher LPS than the LPS involving overlapping partitions. This is done to
allow the static analyzer and code generator to detect an LPS exit within the repeat
loop and place a ghost-region sync stage within the repeat loop.

In addition, this commit does some changes to generate proper C++ expressions for IT
epoch expressions.