fixed a bug in the previous commit
In the case where some segments do not participate in a task execution, the previous code
was letting the non-participating segments to exit without informing the other segments
about their situation. That resulted the participating segments to get stuck in the
communicator setup process indefinitely. This commit fixed that problem.