modified features for reading/writing data parts
This commit is a major step towards enabling reading/writing data files within the environment
management library. This commit also introduces major changes in the structure of the class for
task environments. An unintended consequence of this is reading array dimensions within the
coordinator function from an environment array to a local variable is not working anymore. We
need to fix this problem next.