changed the mechanism of writing environmental data to output files
With the current change, the file writing immediately begins with the occurrance of an
output binding instruction in the coordinator function. Further, the environment management
module takes care of the writing as opposed to executing tasks.

For this change, output-binding should appear after the task that generates the underlying
data executes -- not before it -- as it used to happen up until now.