made changes to keep all LPS data up-to-date after the end of a task execution
In an IT task, there might be multiple allocations for a single data structure because of its
usage in independent LPSes or LPSes having different index ordering requirements. As a result,
depending on the logic of the task, it could happen that some LPS allocations are holding stale
copy of that data at the end of task execution. For example, if Space-A reads variable x then
Space-B writes it and the task end; then Space-A allocation for x will be stale. This commit
added a static analysis of the code that appends dummy reader nodes for stale data structures
in appropriate LPSes so that the structures are updated through synchronization/communication
before the task returns control to the coordinator function.