added some environment manipulation instructions processing codes
With this change, we are back to the place where single-task IT program should run again as we added
processing logic for file I/O on linked and data-parts initialization for created environment variables.
Note that it should theoretically work, but there are significant changes in the previous code generation
process and addition of several new static libraries since the last time we were generating working
executables for IT single-task programs. All these may have many bugs that we have to fix in upcoming
commits.

After we fix all the bugs of the current code, we need to implement the data transfer mechanism for
environmental variables getting assigned from one task to another, i.e., processing of data transfer
and stale-refresh instructions. Then we should be able to run multi-task IT programs.