a temporary patch solution for dynamic memory handling within GPU kernel
For some unknown reason, if we try to assign a shared memory pointer to an index in the
dynamic shared memory pannel by calculating the index through some integer arithmatic on
variables within the kernel, we always get an invalid CUDA memory error. If the same
index calculation is done using constants then the scheme works. We have to investigate
why this is happenning later. For now we keep constant arithmatics in the experimental
project to save time.