added a batch LPU generation routine for GPU PPU controllers
This new routine that internally uses the get-next-LPU routine of older compilers generates
LPUs for different PPU controllers in parallel. So the batch of LPUs that will be offloaded
to the GPU for computation are guaranteed to be intended for different PPUs (SM/Warp) of
the GPU. In some cases we just need enough LPUs to produce work for all PPUs of the GPU --
it does not matter what PPU gets what LPU. In some other cases, however, e.g. LPUs of a sub-
partitioned LPUs needs to be processed by specific PPUs. This new routine provides support
for both modes of execution.