incorporated GPU code exeuctors generation logic in the compiler
GPU code executors execute the specific sub-flows of a task that are intended for offloading.
This commit adds the context specific GPU code executor generation logic in the compiler. The
offload function of the executors, that simulate the task's sub-flow on an specific context,
however, has a blank implementation -- this commit puts in the remaining functionalities of
the executors. We need to discuss the GPU kernel generation strategy first to implement the
offload functions.