partial GPU kernel code generation commit
So far we are able to iterate over the top-most GPU level LPUs that are being staged in by
the host and copy data parts for those LPUs in and out of the shared memory SMs. Later we
have to augment this commit with intra-GPU LPU expansitions for composite stages and code
generation for nested execution stages.

Note that current commits on code generation have many corner cases and efficiency issues
that we have to address in the future.