an experimental change in the communicator
This change impact how location for update or read data from is calculated during interchange between
operating memory data parts and communication buffers of the MPI and segment-local communicators. If
this change does not affect accuracy of the communicators, we can use this as a template for
optimizing communications done for multi-versioned data structures. The current communication mechanism
for such data structure is performing poorly.