two updates related to reduction handling
set up the code to initialize the reduction result variable before a reduction begins and
the code for combining partial results computed by individual PPUs at the end.

Also distinguished between task-global reductions and other reductions. For the former case
only one result variable need to be maintained for the entire task. The latter case will
require dynamic memory allocation and we will handle it in the future.