changed data parts allocation policy for subpartition LPSes
From now on array parts will be allocated in the parent LPS of a subpartition
if the subpartition does not reorder the array parts indexes. This change is
made to keep the data parts count low, which reduces time consumed in initial
sequential setup phase.

If we parallelize the setup phase in the future then we can allocate the parts
in the subpartition LPS. This is a small change in one location only.