an intermediate commit for fixing a memory corruption
The enviornment management module is throwing segmentation fault in the
conjugate gradient program. I am not sure why it is doing so. Further
the problem is only occurring in Rivanna which does not generate any core
file for segmentation faults. So I am quite in loss.

This commit is done hoping the change will improve the situation. Even if
it does not, it should be harmless.