another intermediate commit to fix the current problem with environment management
Again this is related to the conjugate gradient sample program. This commit duplicates
the key of a data part reference when putting the parts lists in the environment as
opposed to use the key directly. It might aleviate the problem that conjugate gradient
is failing for some input size but I am not sure ...