an intermediate commit for parallelization in communicators
This change allow communication buffer preparation before send and buffer
processing after receive to be done cooperatively by all participant threads in
parallel. Only the MPI communication part is done by a single thread.

Currently solely the ghost-region communicator is updated properly to get the
maximum benefit of parallelization.

This change is done in the segmented-memory compiler alone. If it works properly
and improves performance then we will copy the changes in the hybrid compiler.