another intermediate commit on semantic analysis
This code addition determines the dimensionality of arrays acccessed within
a compute stage definition. Note that IT stages are type polymorphic -- not
shape polymorphic. That is, the same stage variable cannot be used once as
a 1D array and as a 2D array afterwards during a different invocation. Such
behaviors would make stage definitions confusing to a programmer.