incorporated a mechanism for front-end/back-end coordination for platform specific code generation
We are keeping the interface definitions of the code-generation functions of the front-end classes
within the front-end compiler. A specific directory has all mock implementation of those functions
so that the front-end compiler can be built separately. The back-end compiler skip the mock
implementation directory; consequently, it is forced to provide its own implementations replacing
the mock implementations.